requirements.txt
django
gunicorn
django-heroku
requests
djangorestframework
markdown
django-filter
django-cors-headers
sklearn
numpy
matplotlib